TraceRA: move trace logging to TraceBuilderPhase.
